Update the player 1 in some benchmark examples
This is the fix to make them cohere to the assumtion that schedulers
should always be able to make moves.  And our experimental proof rule
needs it even more:  The scheduler should be able to a move even when
the input configuration does not make sense (some random combination of
the valid symbols).